@86eq5k2ud comments addressed
